Limit maximum number of strings, memory in JSON input (#1784)
* Limit maximum string count, memory in JSON request

While slightly imprecise due to the simple nature of the parser,
limiting the total number of strings ensures that we have a bounded
number of calls to HMAC within the request body.

This could be improve by not counting strings which are keys in maps, at
the cost of added complexity in the tracking state.

Similarly, by adjusting our use of the tokenizer, we can estimate total
memory usage in the parsed representation of the request body; we want
to limit this to reasonable levels.

This could be improved by adding bucketing overhead for map items,
estimating based on the total number of items in the map.

Co-authored-by: Arran Walker <ajwalker@gitlab.com>
Signed-off-by: Alexander Scheel <ascheel@gitlab.com>

* Apply suggestions from code review

Co-authored-by: wslabosz-reply <w.slabosz@reply.com>
Co-authored-by: Jonas Köhnen <mail@satoqz.net>
Signed-off-by: Alexander Scheel <alexander.m.scheel@gmail.com>

* Use math.MaxInt64 instead of -1, fix tests

This removes the test previously added in the last PR, as we've
incorporated those changes into a different test across layers.

Signed-off-by: Alexander Scheel <ascheel@gitlab.com>

* Update http/json.go

Co-authored-by: Jonas Köhnen <mail@satoqz.net>
Signed-off-by: Alexander Scheel <alexander.m.scheel@gmail.com>

* Update changelog/1756.txt

Co-authored-by: Philipp Stehle <anderschwiedu@gmail.com>
Signed-off-by: Alexander Scheel <alexander.m.scheel@gmail.com>

---------

Signed-off-by: Alexander Scheel <ascheel@gitlab.com>
Signed-off-by: Alexander Scheel <alexander.m.scheel@gmail.com>
Co-authored-by: Arran Walker <ajwalker@gitlab.com>
Co-authored-by: wslabosz-reply <w.slabosz@reply.com>
Co-authored-by: Jonas Köhnen <mail@satoqz.net>
Co-authored-by: Philipp Stehle <anderschwiedu@gmail.com>